requirements.txt
ExifRead==3.0.0
eyed3==0.9.7
matplotlib==3.5.3
# numpy==1.23.5
numpy==1.24.4
# pandas==1.5.3
pandas==2.0.3
# Pillow==11.1.0
Pillow==10.4.0
PyPDF2==3.0.1
seaborn==0.13.2
streamlit==1.40.1
